requirements.txt
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
altair==5.2.0

streamlit==1.29.0

snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1

scikit-learn==1.3.2
xgboost==2.0.3

python-dotenv==1.0.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# Data Science (déjà présents)
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0